# public/api-docs.yaml
openapi: 3.0.0
info:
  title: 'PharmaControl API'
  description: "ERP para farmacias multi-sucursal en México con regulación COFEPRIS.\nAutenticación JWT propio (HS256). Todos los endpoints protegidos requieren Authorization: Bearer <token>."
  version: 1.0.0
servers:
  -
    url: 'http://localhost:8080/api/v1/'
    description: 'Local Docker'
paths:
...
  /api/v1/auth/login:
    post:
      tags:
        - Auth
      summary: 'Iniciar sesión'
      description: 'WEB: tokens en cookies HttpOnly. MOBILE: tokens en body JSON.'
      operationId: d00f4d12443a970455f396756440af3f
      requestBody:
        required: true
        content:
          application/json:
            schema:
              required:
                - email
                - password
                - client_type
              properties:
                email:
                  type: string
                  format: email
                  example: admin@pharmaco.mx
                password:
                  type: string
                  format: password
                  example: 'Ch4ng3M3_N0w!#2026'
                client_type:
                  type: string
                  example: WEB
                  enum: [WEB, MOBILE]
              type: object
      responses:
        '200':
          description: 'Login exitoso'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '423':
          description: 'Cuenta bloqueada'
      security: []
  /v1/auth/logout:
    post:
      tags:
        - Auth
      summary: 'Cerrar sesión'
      description: 'Revoca el token actual, invalida la sesión y borra las cookies HttpOnly.'
      operationId: aa947ef5386e68877a4da98bc98274e2
      responses:
        '200':
          description: 'Sesión cerrada'
        '401':
          $ref: '#/components/responses/Unauthorized'
      security:
        -
          bearerAuth: []
  /v1/auth/refresh:
    post:
      tags:
        - Auth
      summary: 'Renovar tokens'
      description: 'WEB: lee refresh_token de cookie HttpOnly y rota ambas cookies. MOBILE: acepta refresh_token en body y devuelve tokens en body.'
      operationId: 8ec7f584cf069a47d99088337c63343c
      requestBody:
        required: false
        content:
          application/json:
            schema:
              properties:
                refresh_token:
                  description: 'Solo para MOBILE'
                  type: string
              type: object
      responses:
        '200':
          description: 'Tokens renovados'
        '401':
          $ref: '#/components/responses/Unauthorized'
      security: []
...
components:
  schemas:
...
    ErrorResponse:
      properties:
        message:
          type: string
          example: 'Recurso no encontrado.'
        error:
          type: string
          example: NOT_FOUND
          nullable: true
      type: object
    ValidationErrorResponse:
      properties:
        message:
          type: string
          example: 'The name field is required.'
        errors:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
      type: object
...
  responses:
    Unauthorized:
      description: 'Token ausente, inválido o expirado'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            message: Unauthenticated.
...
    UnprocessableEntity:
      description: 'Error de validación'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationErrorResponse'
...
  securitySchemes:
    bearerAuth:
      type: http
      bearerFormat: JWT
      scheme: bearer
...